move minting logic to own component
